[WB-1899.5] Refactor `ButtonCore` to make it an unstyled component (#2590)
## Summary:

Splits `ButtonCore` to keep styles there and creates a new `ButtonUnstyled` to
handle the semantics of the Button HTML elements.

This will allow us to create more "Button" components like the upcoming TB
`ActivityButton` and `DynamicActionButton` that will use the same logic as
`Button` but with different styles.

### Implementation plan:

1. #2579
2. #2580
3. #2581
4. #2589
5. **Refactor `ButtonCore` to make it an unstyled component.(CURRENT)**

Issue: WB-1899

## Test plan:

Verify that the `Button` component still works and looks as expected.

Author: jandrade

Reviewers: jandrade, beaesguerra

Required Reviewers:

Approved By: beaesguerra

Checks: ✅ 12 checks were successful, ⏭️  2 checks have been skipped

Pull Request URL: https://github.com/Khan/wonder-blocks/pull/2590
