+++ .changeset/twelve-paws-kneel.md
@@ -1,0 +1,5 @@
+---
+"@khanacademy/wonder-blocks-button": patch
+---
+
+Splits ButtonCore to keep styles there and creates a new ButtonUnstyled to handle the semantics of the Button HTML elements"
